fix(kube-control-plane): apply patches in legacy kubeadm upgrade
Being that:\n1. We dropped the bind-address configuration from the kubeadm configuration and we move it to kubeadm patches files.\n2. kubeadm legacy (k8s < 1.30) does not support the UpgradeConfiguration object to specify the upgrade commands flags\nWe need to pass explicitly the '--patches' flag to the 'kubeadm upgrade' command